dlp_output/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-dlp==1.0.0
google-cloud-pubsub
requests